requirements.txt
Django==3
django-rq
djangorestframework
